WIP theming for testing (#22)
This is the first part of dark theme enablement, with some incidental linting fixes.

Not everything is appropriately themed. The main thing that I'm aware of is
the Createable that we use for tag entry.

In future, I also want to implement a UI themer, that will enable in-UI customisation

Co-authored-by: James <js@sugrono.com>