Bulk add updates
The bulk update functionality allows operators to add all links on
a page to Jolly Roger.

We leave it to the operator to filter and check for duplicates, so
that they don't add non-puzzle links like navigation, etc.

We now try to get the puzzle title from the anchor text, and only
guess it if it doesn't exist (it's only anchored to an image or
similar). This should yield the best title for puzzles most of the
time.

The main use case for this is adding all puzzles on a round unlock,
but it will be useful to users doing postsolves as well.